+++ order-service/db/migrations/000001_create_table_products.up.sql
@@ -3,7 +3,7 @@ CREATE TABLE products(
     id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
     name VARCHAR(255) NOT NULL,
     price DECIMAL(10, 2) NOT NULL,
-    quantity INT NOT NULL DEFAULT 0 CHECK (quantity >= 0),
+    quantity BIGINT NOT NULL DEFAULT 0 CHECK (quantity >= 0),
     created_at TIMESTAMPTZ NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMPTZ NOT NULL DEFAULT CURRENT_TIMESTAMP
 );
